split up types into there respective files
